feat: refactor modules file to mjs
